Quick fix for faulty dsb-loop
The file read for dsb-calls is really slow... when it works.
When a dsb-sensor is actually present. The long read gives better
resolution and a slow sensor cycle.
When the file is not present, without a sleep in the loop, the cpu is
flooded with synchronous very fast failing io-calls.

Solution: Add sleep to dsb-measurement. This will have to be implemented
better, but later.